Move ViewController Test helpers into a separate file in shared folder
